Generic Language Parsing + Bug Fixes
dictionary changes
- Changed format of pronunciations in the dictionary from CV,VC to
C(V)C.
- Dictionaries must have a "order" parameter as the first line with a
list of single-character characters to check the dictionary for when
writing to file
- Dictionaries will write all words stored into it to file, but any
words that contain characters outside of the ones specified in order may
not be in a logical order.
- Dictionaries are now stored in the "Dictionary" folder
- Now can have multiple dictionaries that are loaded into the parser at
init
* dictionaries are identified by their filename, and are listed in the
language combo box
* initial parsing occurs once the user selects a language. Changing the
language will remove all prior ust changes and redo the parsing based on
the new language's dictionary

parser changes:
- With dictionary changes, all parsing is done through identification of
consonants and vowels, and thus all language specific parsing (for
english) have been removed/updated (except for the english blending of
vowel transitions, which is now a setting)
- User now must select a language and syntax for parsing

Bug fixes:
- Fixed a bug where notes that did not have pitchbend or vibrato was
given the pitchbend and/or vibrato of the note that was originally at
the new note's index
- Fixed a bug where the next note having a pitchbend would cause the
pitchbends of notes beyond the original size of the ust to be incorrect

To fix:
- Opening the parser and clicking "cancel" without doing anything causes
the parser to crash due to no dictionary being selected.